+++ app/src/main/java/com/kuzepa/mydates/domain/repository/EventRepository.kt
@@ -6,8 +6,8 @@ import kotlinx.coroutines.flow.Flow
 
 interface EventRepository {
     suspend fun upsertEvent(event: Event)
-    suspend fun getEventById(id: Int): Event?
-    suspend fun deleteEventById(id: Int)
+    suspend fun getEventById(id: Long): Event?
+    suspend fun deleteEventById(id: Long)
     fun getAllEvents(sortOption: SortOption = SortOption.UNSPECIFIED): Flow<List<Event>>
     fun getEventsByMonth(month: Int, sortOption: SortOption): Flow<List<Event>>
 }
